# backend/openapi.json
{"openapi":"3.1.0","info":{"title":"VulnScanner API","description":"Enhanced Vulnerability Scanner API","version":"1.0.0"},"paths":{"/api/auth/login":{"post":{"tags":["local-auth"],"summary":"Login","description":"Local login for SQLite development.\nAuthenticates against the local SQLite database.","operationId":"login_api_auth_login_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserLogin"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Token"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/auth/user":{"get":{"tags":["local-auth"],"summary":"Get Current User Info","description":"Get current user information.","operationId":"get_current_user_info_api_auth_user_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserPublic"}}}}},"security":[{"HTTPBearer":[]}]}},"/api/auth/logout":{"post":{"tags":["local-auth"],"summary":"Logout","description":"Logout endpoint (for compatibility).","operationId":"logout_api_auth_logout_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/payments/checkout":{"post":{"tags":["payments"],"summary":"Checkout","operationId":"checkout_api_payments_checkout_post","parameters":[{"name":"plan_id","in":"query","required":true,"schema":{"type":"string","title":"Plan Id"}},{"name":"x-csrf-token","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"X-Csrf-Token"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/projects":{"get":{"tags":["dashboard"],"summary":"List Projects","description":"List all projects for the current user.","operationId":"list_projects_api_v1_projects_get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":1000,"minimum":1,"default":50,"title":"Limit"}},{"name":"offset","in":"query","required":false,"schema":{"type":"integer","minimum":0,"default":0,"title":"Offset"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ProjectResponse"},"title":"Response List Projects Api V1 Projects Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"tags":["dashboard"],"summary":"Create Project","description":"Create a new project.","operationId":"create_project_api_v1_projects_post","security":[{"HTTPBearer":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProjectCreate"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProjectResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/projects/{project_id}":{"get":{"tags":["dashboard"],"summary":"Get Project","description":"Get a specific project by ID.","operationId":"get_project_api_v1_projects__project_id__get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"project_id","in":"path","required":true,"schema":{"type":"integer","title":"Project Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProjectResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"tags":["dashboard"],"summary":"Update Project","description":"Update a project.","operationId":"update_project_api_v1_projects__project_id__put","security":[{"HTTPBearer":[]}],"parameters":[{"name":"project_id","in":"path","required":true,"schema":{"type":"integer","title":"Project Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProjectUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProjectResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["dashboard"],"summary":"Delete Project","description":"Delete a project.","operationId":"delete_project_api_v1_projects__project_id__delete","security":[{"HTTPBearer":[]}],"parameters":[{"name":"project_id","in":"path","required":true,"schema":{"type":"integer","title":"Project Id"}}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/projects/{project_id}/summary":{"get":{"tags":["dashboard"],"summary":"Get Project Summary","description":"Get project summary statistics.","operationId":"get_project_summary_api_v1_projects__project_id__summary_get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"project_id","in":"path","required":true,"schema":{"type":"integer","title":"Project Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProjectSummary"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/projects/{project_id}/scans":{"get":{"tags":["dashboard"],"summary":"List Project Scans","description":"List scan sessions for a project.","operationId":"list_project_scans_api_v1_projects__project_id__scans_get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"project_id","in":"path","required":true,"schema":{"type":"integer","title":"Project Id"}},{"name":"status","in":"query","required":false,"schema":{"anyOf":[{"$ref":"#/components/schemas/ScanStatus"},{"type":"null"}],"title":"Status"}},{"name":"start_date","in":"query","required":false,"schema":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Start Date"}},{"name":"end_date","in":"query","required":false,"schema":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"End Date"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":1000,"minimum":1,"default":50,"title":"Limit"}},{"name":"offset","in":"query","required":false,"schema":{"type":"integer","minimum":0,"default":0,"title":"Offset"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ScanSessionResponse"},"title":"Response List Project Scans Api V1 Projects  Project Id  Scans Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"tags":["dashboard"],"summary":"Create Scan Session","description":"Create a new scan session.","operationId":"create_scan_session_api_v1_projects__project_id__scans_post","security":[{"HTTPBearer":[]}],"parameters":[{"name":"project_id","in":"path","required":true,"schema":{"type":"integer","title":"Project Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ScanSessionCreate"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ScanSessionResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/scans/{scan_id}":{"get":{"tags":["dashboard"],"summary":"Get Scan Session","description":"Get a specific scan session.","operationId":"get_scan_session_api_v1_scans__scan_id__get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"scan_id","in":"path","required":true,"schema":{"type":"integer","title":"Scan Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ScanSessionResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/overview":{"get":{"tags":["dashboard"],"summary":"Get Dashboard Overview","description":"Get comprehensive dashboard overview with real-time metrics and enhanced analytics.","operationId":"get_dashboard_overview_api_v1_overview_get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"time_range","in":"query","required":false,"schema":{"type":"string","pattern":"^(1h|6h|24h|7d|30d|90d)$","default":"7d","title":"Time Range"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DashboardOverview"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/metrics/live":{"get":{"tags":["dashboard"],"summary":"Get Live Metrics","description":"Get live dashboard metrics with real-time updates.","operationId":"get_live_metrics_api_v1_metrics_live_get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"metric_type","in":"query","required":false,"schema":{"type":"string","pattern":"^(all|scans|urls|forms|performance)$","default":"all","title":"Metric Type"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","additionalProperties":true,"title":"Response Get Live Metrics Api V1 Metrics Live Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/metrics/update":{"post":{"tags":["dashboard"],"summary":"Update Dashboard Metric","description":"Update a dashboard metric and broadcast to all connected clients.","operationId":"update_dashboard_metric_api_v1_metrics_update_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DashboardMetricUpdate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/api/v1/realtime/updates":{"get":{"tags":["dashboard"],"summary":"Get Realtime Updates","description":"Get recent real-time updates for the user.","operationId":"get_realtime_updates_api_v1_realtime_updates_get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":100,"minimum":1,"default":50,"title":"Limit"}},{"name":"since","in":"query","required":false,"schema":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Since"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","additionalProperties":true,"title":"Response Get Realtime Updates Api V1 Realtime Updates Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/realtime/broadcast":{"post":{"tags":["dashboard"],"summary":"Broadcast Custom Update","description":"Broadcast a custom real-time update to all connected clients.","operationId":"broadcast_custom_update_api_v1_realtime_broadcast_post","requestBody":{"content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Update Data"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/api/v1/analytics/summary":{"get":{"tags":["dashboard"],"summary":"Get Analytics Summary","description":"Get comprehensive analytics summary for the dashboard.","operationId":"get_analytics_summary_api_v1_analytics_summary_get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"time_range","in":"query","required":false,"schema":{"type":"string","pattern":"^(1h|6h|24h|7d|30d|90d)$","default":"7d","title":"Time Range"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","additionalProperties":true,"title":"Response Get Analytics Summary Api V1 Analytics Summary Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/dashboard/projects-summary":{"get":{"tags":["dashboard"],"summary":"Get Projects Summary","description":"Get projects summary for dashboard display.","operationId":"get_projects_summary_api_v1_dashboard_projects_summary_get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":100,"minimum":1,"default":10,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"type":"object","additionalProperties":true},"title":"Response Get Projects Summary Api V1 Dashboard Projects Summary Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/dashboard/recent-activity":{"get":{"tags":["dashboard"],"summary":"Get Recent Activity","description":"Get recent activity for dashboard display.","operationId":"get_recent_activity_api_v1_dashboard_recent_activity_get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":50,"minimum":1,"default":10,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"type":"object","additionalProperties":true},"title":"Response Get Recent Activity Api V1 Dashboard Recent Activity Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/dashboard/vulnerability-trends":{"get":{"tags":["dashboard"],"summary":"Get Vulnerability Trends","description":"Get vulnerability trends for dashboard charts.","operationId":"get_vulnerability_trends_api_v1_dashboard_vulnerability_trends_get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"days","in":"query","required":false,"schema":{"type":"integer","maximum":365,"minimum":1,"default":30,"title":"Days"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"type":"object","additionalProperties":true},"title":"Response Get Vulnerability Trends Api V1 Dashboard Vulnerability Trends Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/dashboard/scan-statistics":{"get":{"tags":["dashboard"],"summary":"Get Scan Statistics","description":"Get scan statistics for dashboard display.","operationId":"get_scan_statistics_api_v1_dashboard_scan_statistics_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Response Get Scan Statistics Api V1 Dashboard Scan Statistics Get"}}}}},"security":[{"HTTPBearer":[]}]}},"/api/v1/health":{"get":{"tags":["dashboard"],"summary":"Dashboard Health Check","description":"Health check endpoint for dashboard services.","operationId":"dashboard_health_check_api_v1_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Response Dashboard Health Check Api V1 Health Get"}}}}},"security":[{"HTTPBearer":[]}]}},"/api/health":{"get":{"summary":"Health","description":"Health check endpoint.","operationId":"health_api_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"DashboardMetricUpdate":{"properties":{"value":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Value"},"metadata":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Metadata"}},"type":"object","title":"DashboardMetricUpdate","description":"Schema for updating dashboard metrics."},"DashboardOverview":{"properties":{"total_projects":{"type":"integer","title":"Total Projects"},"total_scans":{"type":"integer","title":"Total Scans"},"active_scans":{"type":"integer","title":"Active Scans"},"total_urls_discovered":{"type":"integer","title":"Total Urls Discovered"},"total_forms_found":{"type":"integer","title":"Total Forms Found"},"total_technologies_detected":{"type":"integer","title":"Total Technologies Detected"},"recent_projects":{"items":{"$ref":"#/components/schemas/ProjectSummary"},"type":"array","title":"Recent Projects"},"recent_scans":{"items":{"$ref":"#/components/schemas/ScanSummary"},"type":"array","title":"Recent Scans"}},"type":"object","required":["total_projects","total_scans","active_scans","total_urls_discovered","total_forms_found","total_technologies_detected","recent_projects","recent_scans"],"title":"DashboardOverview","description":"Dashboard overview data."},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ProjectCreate":{"properties":{"name":{"type":"string","maxLength":100,"minLength":1,"title":"Name","description":"Project name"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description","description":"Project description"},"target_domain":{"type":"string","maxLength":255,"minLength":1,"title":"Target Domain","description":"Target domain to scan"},"scope_rules":{"items":{"type":"string"},"type":"array","title":"Scope Rules","description":"URL scope rules (regex patterns)"}},"type":"object","required":["name","target_domain"],"title":"ProjectCreate","description":"Schema for creating a new project."},"ProjectResponse":{"properties":{"name":{"type":"string","maxLength":100,"minLength":1,"title":"Name","description":"Project name"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description","description":"Project description"},"target_domain":{"type":"string","maxLength":255,"minLength":1,"title":"Target Domain","description":"Target domain to scan"},"scope_rules":{"items":{"type":"string"},"type":"array","title":"Scope Rules","description":"URL scope rules (regex patterns)"},"id":{"type":"integer","title":"Id"},"owner_id":{"type":"integer","title":"Owner Id"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"updated_at":{"type":"string","format":"date-time","title":"Updated At"}},"type":"object","required":["name","target_domain","id","owner_id","created_at","updated_at"],"title":"ProjectResponse","description":"Schema for project response."},"ProjectSummary":{"properties":{"id":{"type":"integer","title":"Id"},"name":{"type":"string","title":"Name"},"target_domain":{"type":"string","title":"Target Domain"},"total_scans":{"type":"integer","title":"Total Scans"},"last_scan_date":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Last Scan Date"},"total_urls_discovered":{"type":"integer","title":"Total Urls Discovered"},"total_forms_found":{"type":"integer","title":"Total Forms Found"}},"type":"object","required":["id","name","target_domain","total_scans","last_scan_date","total_urls_discovered","total_forms_found"],"title":"ProjectSummary","description":"Summary information for a project."},"ProjectUpdate":{"properties":{"name":{"anyOf":[{"type":"string","maxLength":100,"minLength":1},{"type":"null"}],"title":"Name"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"target_domain":{"anyOf":[{"type":"string","maxLength":255,"minLength":1},{"type":"null"}],"title":"Target Domain"},"scope_rules":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Scope Rules"}},"type":"object","title":"ProjectUpdate","description":"Schema for updating a project."},"ScanConfigurationSchema":{"properties":{"max_depth":{"type":"integer","maximum":10.0,"minimum":1.0,"title":"Max Depth","description":"Maximum crawl depth","default":3},"max_pages":{"type":"integer","maximum":100000.0,"minimum":10.0,"title":"Max Pages","description":"Maximum pages to crawl","default":1000},"requests_per_second":{"type":"integer","maximum":100.0,"minimum":1.0,"title":"Requests Per Second","description":"Request rate limit","default":10},"timeout":{"type":"integer","maximum":300.0,"minimum":5.0,"title":"Timeout","description":"Request timeout in seconds","default":30},"follow_redirects":{"type":"boolean","title":"Follow Redirects","description":"Follow HTTP redirects","default":true},"respect_robots":{"type":"boolean","title":"Respect Robots","description":"Respect robots.txt","default":true},"user_agent":{"type":"string","title":"User Agent","description":"User agent string","default":"Enhanced-Vulnerability-Scanner/1.0"},"scope_patterns":{"items":{"type":"string"},"type":"array","title":"Scope Patterns","description":"URL scope patterns"},"exclude_patterns":{"items":{"type":"string"},"type":"array","title":"Exclude Patterns","description":"URL exclude patterns"},"authentication":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Authentication","description":"Authentication configuration"}},"type":"object","title":"ScanConfigurationSchema","description":"Schema for scan configuration."},"ScanSessionCreate":{"properties":{"configuration":{"$ref":"#/components/schemas/ScanConfigurationSchema"}},"type":"object","required":["configuration"],"title":"ScanSessionCreate","description":"Schema for creating a new scan session."},"ScanSessionResponse":{"properties":{"id":{"type":"integer","title":"Id"},"project_id":{"type":"integer","title":"Project Id"},"status":{"$ref":"#/components/schemas/ScanStatus"},"start_time":{"type":"string","format":"date-time","title":"Start Time"},"end_time":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"End Time"},"configuration":{"additionalProperties":true,"type":"object","title":"Configuration"},"stats":{"additionalProperties":true,"type":"object","title":"Stats"},"created_by":{"type":"integer","title":"Created By"}},"type":"object","required":["id","project_id","status","start_time","end_time","configuration","stats","created_by"],"title":"ScanSessionResponse","description":"Schema for scan session response."},"ScanStatus":{"type":"string","enum":["pending","running","completed","failed","cancelled"],"title":"ScanStatus","description":"Scan status enumeration."},"ScanSummary":{"properties":{"id":{"type":"integer","title":"Id"},"project_id":{"type":"integer","title":"Project Id"},"status":{"$ref":"#/components/schemas/ScanStatus"},"start_time":{"type":"string","format":"date-time","title":"Start Time"},"end_time":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"End Time"},"urls_discovered":{"type":"integer","title":"Urls Discovered"},"forms_extracted":{"type":"integer","title":"Forms Extracted"},"technologies_detected":{"type":"integer","title":"Technologies Detected"}},"type":"object","required":["id","project_id","status","start_time","end_time","urls_discovered","forms_extracted","technologies_detected"],"title":"ScanSummary","description":"Summary information for a scan session."},"Token":{"properties":{"access_token":{"type":"string","title":"Access Token"},"token_type":{"type":"string","title":"Token Type","default":"bearer"},"expires_in":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Expires In"},"refresh_token":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Refresh Token"}},"type":"object","required":["access_token"],"title":"Token","description":"JWT token response model."},"UserLogin":{"properties":{"email":{"type":"string","format":"email","title":"Email"},"password":{"type":"string","title":"Password"}},"type":"object","required":["email","password"],"title":"UserLogin","description":"User login model."},"UserPublic":{"properties":{"id":{"type":"string","title":"Id"},"email":{"type":"string","format":"email","title":"Email"},"full_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Full Name"},"email_confirmed":{"type":"boolean","title":"Email Confirmed","default":false},"created_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Created At"},"updated_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Updated At"}},"type":"object","required":["id","email"],"title":"UserPublic","description":"Public user information model."},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"HTTPBearer":{"type":"http","scheme":"bearer"}}}}